filter/packet: remove nop frame.
Filtering is no longer inlined in the main program, but it is now part
of a subprog. This has multiple benefits:

- stack separation
- reduced risk of unintentional register clobbering
- potentially reduced load time

Given the removal of the nop trick, pseudo calls are now fixed up
while scanning the program, also func_info and line_info offsets are
adjusted accordingly.
All the instructions need to be checked as relative jumps may point to
negative offset and so preceding subprogs, this is a bit tedious but
doesn't add much overhead compared to the previous way as the only
additional instructions scanned would be the hooks stubs.

This whole change also paves the way for dynamically generate
meta-filtering.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>